Update readme instructions
Running the commands in the readme results in the server starting on
port 3000 and the client expects it to be on 8080, so we get the error:

> Doh - disconnected - no snek running at localhost:8080

I've updated the instructions to set the port for the server, there is also a
Procfile which the instructions don't use, it could be a good idea to have a
single Procfile in the base directory that starts all 3 processes.

Also the client doesn't use node, so no need to install packages via yarn.